{{cookiecutter.project_slug}}/requirements.txt
Django==2.2.5
djangorestframework==3.10.3
django-cors-headers==3.1.0
django-filter==2.2.0
coreapi==2.3.3
factory-boy==2.12.0
django-anymail==6.1.0
Pillow==6.1.0
djangorestframework-filters==0.11.1
python-decouple==3.1
django-storages==1.7.2
dj_database_url==0.5.0
zappa==0.48.2
psycopg2==2.7.5
jsonfield==2.0.2
django-model-utils==3.2.0